all is working!
This completes my goal for this step in work - we wanted
to create an example using fluxion that parsed topology
information into a graph. Likely we next want to add
this to fluence, and maybe have some kind of merge
of this JGF with our nodes? Not sure, but sounds fun.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>